support hooks on websocket
