+++ lib/rule_default.js
@@ -78,4 +78,28 @@ module.exports = {
   *onClientSocketError(requestDetail, error) {
     return null;
   },
+
+  /**
+   *
+   * @param {object} messageDetail
+   * @param {string|buffer} messageDetail.data
+   * @param {string} messageDetail.url
+   * @param {object} messageDetail.requestOptions
+   * @returns
+   */
+  *beforeSendWsMessageToServer(messageDetail) {
+    return null;
+  },
+
+  /**
+   *
+   * @param {object} messageDetail
+   * @param {string|buffer} messageDetail.data
+   * @param {string} messageDetail.url
+   * @param {object} messageDetail.requestOptions
+   * @returns
+   */
+  *beforeSendWsMessageToClient(messageDetail) {
+    return null;
+  }
 };
